File structure update: 1 download for everything
This folder contains all the files needed to run the Cornell Culvert Model, including the CulvertStartKit and CulvertEvaluation files.  The CulvertModelInstructions have been updated to include the modified file structure, as well as some slight modifications to the python code that allow files to be saved directly to the data folder so the user doesn't have to manually move them after creation.